Fixed conflict, added weapon drops

+++ ReaversFPS/Assets/Scripts/PickUps/Weapons/gunStats.cs
@@ -9,7 +9,7 @@ public class gunStats : ScriptableObject
     public int ammoCount;
     public int magazineCount;
     public int damage;
-    public int fireRate;
+    public float fireRate;
     public int shootDistance;
     public AudioClip audioClip;
     public GameObject model;
